Move test to superclass
One of the previous bug fixes must have addressed this

commit-id: I0b79e8c1
